MA-5979 - [Android_MX6SX/SL_SD/EVK]USB Camera: Audio plays with frozen frame
when play to the end of recorded video clip. 100%

In MediaCodecSource.cpp, when has input data, push it's time stamp to a list.
When has output data, erase a stamp from the list.
But if the frame is dropped in h264 encoder, the list will enlarge continuously.
Thus video time stamp drift and memory leak.

To fix it, not use a list, just use time stamp got from output.

Signed-off-by: Fang Hui <b31070@freescale.com>