MA-7487 AudioTrack::getTimestamp always failed when play videos on sabreauto_dl board
on sabreauto board, if play media stream with multi-channel audio(5.1/7.1),
omx's AudioSink->getTimestamp() returns -38

Find that if switches to 6 channel audio, due to audio_policy.conf's
channel config, direct output thread is created. The thread->mOutput->flags
is compatible with direct output’s flag, but the mFlags in AudioTrack,
which is set at the creation of AudioTrack(af->createTrack), is 0,
this lead to isDirect() == 0, so direct output(here is esai)’s gettimestamp
can’t be called.

Fixed by send the trackFlags with DIRECT flag setted in createTrack_l() of
direct thread back to AudioTrack, and set the corresponding mFlags.

Signed-off-by: Winter Wang <b55847@freescale.com>